New branch "new-nav-system" created to replace the NavMesh
Created the new branch "new-nav-system" to replace the use of the NavMesh with a different navigation system to move Bob from its start position to the exit position. This to avoid the complexities and great load times of the NavMesh. All the following changes were done in the LevelManager class:

- Changed the name of the "IsPathToDestFound" function to "IsPathToExitFound".
- Added the new variable pathToExit (List) to store the path to the exit and modified the IsPathToExitFound function to store the coordinates ("edges") of the path to the exit once it is found.
- Added new function "PrintPathToExit" to print (Debug.Log) the pathToExit list (for debugging purposes).