DOTween module settings changed
Committing this change on the settings file of the DOTween module as it was not taken by my previous commit. This setting was made when the DOTween plugin was installed, the Physics and Physics2D modules of the plugin were disabled as they were not going to be used in the project.